chore: add nav utils
