+++ src/App.tsx
@@ -11,9 +11,13 @@ import Navigator from './modules/Navigation'
 import createStore from './redux/createStore'
 
 import { PersistGate } from 'redux-persist/integration/react'
+import { onNavigationStateChange } from './modules/Navigation/Navigation.util'
 
 if (HIDE_WARNINGS) {
   console.disableYellowBox = true // tslint:disable-line
+}
+if (__DEV__) {
+  console.ignoredYellowBox = ['react-native-i18n module is not correctly link'] // tslint:disable-line
 }
 
 const { store, persistor } = createStore()
@@ -37,7 +41,7 @@ const App = () => (
     <ThemeProvider theme={theme}>
       <PersistGate loading={null} persistor={persistor}>
         <ApolloProvider client={client}>
-          <Navigator />
+          <Navigator onNavigationStateChange={onNavigationStateChange} />
         </ApolloProvider>
       </PersistGate>
     </ThemeProvider>
